[fully_async] address rollout failure review feedback
Move rollout failure signaling helpers into detach_utils per review.

Let the trainer treat rollout error signals as termination while rollouter.fit continues raising the original failure to main.

Keep CPU coverage for signal formatting and task failure detection.
